Hardlink tracker increment / decrement didn't allow or the scenario where >50% of hardlink legs are moved from one directory A to another directory B, but A is traversed during merged before B is. The result is you would get a higher number for nlink than actually there would be legs for.

The solution here is to allow nlink to go negative, and only remove zero-ed hardlink entries at the very end of the merge.

Testing Done:
New and existing tests

Reviewed at http://gut/review/r/1653/

+++ daemon/refresh.go
@@ -329,14 +329,14 @@ func (wsr *WorkspaceRoot) refreshRemoteHardlink_(c *ctx,
 	if entry, exists := wsr.hardlinkTable.hardlinks[id]; !exists {
 		c.vlog("Adding new hardlink entry with id %d", id)
 		newLink := newLinkEntry(hardlink.Record())
-		newLink.nlink = hardlink.Nlinks()
+		newLink.nlink = int64(hardlink.Nlinks())
 		wsr.hardlinkTable.hardlinks[id] = newLink
 	} else {
 		c.vlog("found mapping %d -> %s (nlink %d vs. %d)", id,
 			entry.record.Filename(), hardlink.Nlinks(), entry.nlink)
 		oldRecord := entry.record
 
-		entry.nlink = hardlink.Nlinks()
+		entry.nlink = int64(hardlink.Nlinks())
 		entry.record = hardlink.Record()
 		wsr.hardlinkTable.hardlinks[id] = entry
